requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
torchaudio>=2.0.0
einops>=0.6.0

# Graph Neural Networks
torch-geometric>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Spatial Transcriptomics
scanpy>=1.9.0
anndata>=0.9.0
squidpy>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
# scikit-image>=0.21.0

# File I/O
h5py>=3.9.0
zarr>=2.16.0
pyyaml>=6.0

# Web Interface
gradio>=4.0.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
tensorboard>=2.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0

# Image IO
tifffile>=2023.0.0
imagecodecs>=2023.0.0
